+++ modules/hanna-react/src/utils.ts
@@ -1,5 +1,6 @@
 export * from './utils/browserSide';
 export * from './utils/config';
+export * from './utils/HannaUIState';
 export * from './utils/useDidChange';
 export * from './utils/useFormatMonitor';
 export * from './utils/useGetSVGtext';